requirements.txt
Django>=5.0,<6.0
djangorestframework
python-dotenv==1.1.0
django-environ==0.12.0
django-tailwind
celery
django-celery-beat
python-decouple==3.8
python-docx==1.1.2
pymysql==1.1.1
PyPDF2==3.0.1
pydantic==2.11.2
gunicorn==23.0.0
stripe
openai
Pillow==11.1.0
whitenoise==6.9.0
django-stripe-payments==2.0.0
django-tinymce==4.1.0